Make the release workflow more resilient
## Summary

Currently, it is possible to create a tag and then have the release fail. Since we can't edit the tag (https://github.com/charliermarsh/ruff/issues/4468), this reworks the release so that the tag is created inside the release workflow. This leaves as a failure mode that we have published to pypi but then creating the tag or GitHub release doesn't work, but in this case we can restart and the pypi upload is just skipped because we use the skip existing option.

The release workflow is started by a workflow dispatch with the tag instead of creating the tag yourself. You can start the release workflow without a tag to do a test run.

This also adds docs on how to release and a small style improvement for the maturin integration.

## Test Plan

Normal CI, thorough review and hoping that the next release will work